Tarefa1

Documentation

generateMaze

Arguments

:: Int
Variável da largura

-> Int
Variável do comprimento

-> Int
Seed do labririnto

-> Maze
Labririnto

Introdução

Nesta Tarefa 1 tinhamos como objetivo gerar labirintos aleatórios para o jogo do Pacman.

Objetivos e Desenvolvimento

Para a criação destes labirintos tinhamos de obedecer a certas regras tais como: o labrinto ser rodeado por paredes exceto no túnel e este túnel deveria ser a faixa mais central do labirinto. Para além disto, todos os corredores deveriam ter o mesmo comprimento e a casa dos fantasmas deveria estar localizada no meio do labrinito.

Para a construção dos labrintos, de forma a construir os labirintos corretamente, tivemos de construir duas casas de fantasmas pois de acordo com a altura do labririnto, caso esta fosse par ou ímpar teria de ser colocada de forma diferente.

Discussão e Conlusão

Sendo esta a primeira tarefa e o primeiro contacto com a formação do jogo Pacman, este deixou-nos bastante empenhados e com uma visão diferente na estruturação de um jogo.

Gera o Labririnto. (Função Principal)

limMaze

Arguments

:: Int
Variável da largura

-> Int
Variável do comprimento

-> Int
Seed do labrinto

-> Maze
Labririnto

Junta as funções relativas às paredes de forma a criar um labrinto correto.

walls

Arguments

:: Int
Variável da largura

-> Corridor
Lista de paredes

Altera as listas de forma a criar as Paredes superior e infeior.

walls2

Arguments

:: Int
Variável da altura

-> Maze
Labirinto já criado

-> Maze
Labirinto com as paredes laterais corretas e túnel

Altera as listas de forma a criar as paredes laterais e o túnel.

putHouse

Arguments

:: Int
Variável da largura

-> Int
Variável do comprimento

-> Maze
Labrinto já construído

-> Maze
Labrinto modificado com a casa dos fantasmas

Define o local da casa dos fantasmas no labririnto.

escolheParte

Arguments

:: Int
Variável da largura

-> Int
Variável do comprimento

-> Corridor
Corredor que será substituído

-> Corridor
Corredor substituído pelo corredor correspondente da casa dos fantasmas

Introduz a Casa dos Fantasmas no Labirinto.

ghostHouse

Arguments

:: Int
Variável da largura

-> Maze
Labirinto com a casa dos fantasmas

Peças da casa dos Fantasmas caso a largura do labririnto seja par ou ímpar.

genMaze

Arguments

:: Int
Variável da Largura

-> Int
Variável do comprimento

-> Int
Variável da seed

-> Maze
Labirinto gerado

Função que gera o Labirinto.

nrAleatorio :: Int -> Int

Dado uma seed, retorna um inteiro gerado aleatóriamente.

geraAl :: Int -> Int -> [Int]

Com a seed fornecida, retorna uma lista de n inteiros gerados aleatóriamente.

subLista :: Int -> [a] -> [[a]]

Converte uma lista em uma lista de listas de tamanho n.

convMaze :: [[Int]] -> Maze

Junta todas as listas de forma a criar um Labririnto

convCorredor :: [Int] -> Corridor

Converte uma lista de inteiros num corredor

convPiece :: Int -> Piece

Converte um número inteiro numa peça

printMaze' :: Maze -> IO ()

Converte um labirinto em IO()

mazeToString :: Maze -> String

Transforma o Labirinto criado em uma string

corridorToString :: Corridor -> String

Converte um Corredor em uma string