Tarefa2

Documentation

play

Arguments

:: Play
Jogada a ser realizada

-> State
Estado do jogo antes do movimento do jogador

-> State
Estado de jogo após o movimento do jogador

Introdução

Esta Tarefa 2 tinhamos como objetivo implementar a opção de efetuar jogadas.

Objetivos e Desenvolvimento

Esta tarefa é uma das mais importantes deste jogo pois é a partir dela que podemos movimentar o pacman pelo labririnto, perder vidas, passar para o modo normal, Mega ou Dying do pacman, trocar o modo dos fanstasmas (seja ele vivo ou morto), impossibilitar a entrada do pacman na casa dos fantasmas, o pacman ter a possibilidade de comer os fantasmas e estes regressarem à sua casa, adicionar pontuação caso o pacman coma comida pequena ou grande e/ou fantasmas e atualizar o estado do labrinto apgando por exemplo a comida que o pacman comeu.

Discussão e Conlusão

Esta tarefa foi sem dúvida a mais trabalhosa no entanto recompensadora porque a partir deste momento é possível movimentar o pacman pelo labririnto.

Realiza o movimento do jogador no labririnto. (Função Principal)

oldplayer

Arguments

:: Int
Número identificador do player

-> [Player]
Lista de jogadores

-> Player
Jogador obtido

Recebe o número identificador do player e devolve o player correspondente.

setPlayerOri :: Player -> Orientation -> Player

Recebe um jogador, uma orientação e devolve um jogador com a orientação fornecida.

newList

Arguments

:: Player
Jogador modificado

-> [Player]
Lista dos jogadores

-> [Player]
Lista dos jogadores atualizada

Substitui o jogador na lista de jogadores.

oripl :: Player -> Orientation

Recebe um jogador e devolve a sua orientação.

nextPiece

Arguments

:: Maze
Labirinto

-> Coords
Coordenadas

-> Orientation
Orientação

-> Piece
Peça com as coordenadas atualizadas

Devolve a peça que se encontra na coordenada modificada pela orientação fornecida.

nonextPiece

Arguments

:: Maze
Labririnto

-> Coords
Coordenadas

-> Orientation
Orientação

-> Bool
Verdadeiro ou Falso

Verifica se, de acordo com as coordenadas e a orientação fornecida, existe peças nesse local.

coordTunel

Arguments

:: Coords
Coordenadas de um lado do tunel

-> Maze
Labririnto

-> Coords
Coordenadas do lado oposto do tunel

Troca as coordenadas de um lado tunel para o outro

mudaCoord :: Coords -> Orientation -> Coords

Redefine as coordenadas de acordo com a orientação.

coordspl :: Player -> Coords

Recebe um jogador e devolve as suas coordenadas.

putCoords :: Player -> Coords -> Player

Recebe um jogador, as novas coordenadas e devolve o jogador com as coordenadas atualizadas.

deletePiece

Arguments

:: Maze
Labirinto antes da alteração

-> Coords
Coordenada da peça a ser substituída por Empty

-> Maze
Labririnto atualizado

Elimina a peça nas coodenadas fornecidas.

sameOriGhost

Arguments

:: [Player]
Lista dos jogadores

-> Coords
Possíveis coordenadas do fantasma

-> Bool
Verdadeiro ou Falso

Verifica a existência de um fantasma em determinadas coordenadas.

sameOriPacman

Arguments

:: [Player]
Lista dos jogadores

-> Coords
Possíveis coordenadas do pacman

-> Bool
Verdadeiro ou Falso

Verifica se existe um pacman numa determinada coordenada.

selectOriplayer

Arguments

:: [Player]
Lista dos jogadores

-> Coords
Coordenadas da possível localização

-> Player
Fantasma seleccionado

Seleciona o fantasma da lista dos jogadores nas coordenadas fornecidas.

singlePac :: [Player] -> Player

Recebe a lista de jogadores e devolve o pacman.

beMega :: Player -> Bool

Verifica se o estado do pacman é Mega.

ifNotMegaSetAlive :: [Player] -> [Player]

Recebe a lista de jogadores e caso não exista nenhum pacman no estado Mega, altera todos os estados dos ghosts para Alive.

typeghost :: Player -> GhostMode

Recebe um Fantasma e devolve o seu modo.

bePacman :: Player -> Bool

Testa se o player é um pacman

addscore :: Int -> Player -> Player

Recebe um número inteiro que será adiconado à pontuação do jogador selecionado.

setmega :: Player -> Player

Recebe um pacman e altera o seu estado para Mega.

lessTimeMega :: Player -> Player

Diminui o tempo do pacman caso esteja em Mega (velocidade 1)

lesslife :: Player -> Player

Recebe um jogador ao qual será retirada uma vida e, caso tenha 0 vidas altera os seu estado para Dying.

halfspeed :: [Player] -> [Player]

Recebe a lista de jogadores e reduz para metade a velocidade de todos os fantasmas.

doublespeed :: Player -> Player

Recebe um fantasma e duplica a sua velocidade.

allghostAlive :: [Player] -> [Player]

Recebe a lista de jogadores e altera o estado de todos os fantasmas para Alive.

allghostdead :: [Player] -> [Player]

Recebe a lista de jogadores e altera o estado de todos os fantasmas para Dead.

invertOri :: Orientation -> Orientation

Recebe uma orientação e devolve-a invertida.

aliveghost :: Player -> Player

Recebe um fantasma e altera o seu estado para Alive.

openCloseMouth :: Player -> Player

Abre e fecha a boca do pacman