Tarefa3

Documentation

compactMaze :: Maze -> Instructions

Introdução

Esta Tarefa 3 tínhamos como objetivo compactar o labririnto.

Objetivos e Desenvolvimento

Nesta tarefa 3 criamos funções que analisavam o labirinto feito e o colocavam na forma de pares em que o primeiro elemento seria no número de vezes que essa peça seria repetida num determinado corredor e o segundo elemento a peça.

Para completar esta compactação, caso houvesse corredores idênticos estes seriam mostrados como Repeat x, sendo este x o número da linha anterior com um corredor igual.

Discussão e Conlusão

Com a finalização desta tarefa somos capaz de compactar qualquer labirinto eficazmente e desta forma fazer com que este ocupe menos espaço.

mazeSmall

Arguments

:: [[(Int, Piece)]]
Labirinto fixo

-> [[(Int, Piece)]]
Labirinto alterável

-> [Instruction]
Labirinto comprimido

O primeiro labirinto é um valor fixo que serve apenas para comparar com o segundo labirindo para que depois da primeira volta, a função recursiva ainda tenha como se comparar ao labirindo inicial.

nSame

Arguments

:: [(Int, Piece)]
Corredor comprimido

-> [[(Int, Piece)]]
Conjunto de corredores comprimidos (labirinto)

-> Int
Número de corredores iguais

Calcula quantos corredores iguais ao corredor selecionado existem no labirinto.

fstCorridor

Arguments

:: [(Int, Piece)]
Corredor comprimido

-> [[(Int, Piece)]]
Conjunto de corredores comprimidos (labirinto)

-> Int
Indice do primeiro

Calcula o indice do primeiro corredor igual ao corredor selecionada dentro do labirinto.

mazeNormal :: Maze -> [[(Int, Piece)]]

Transforma um labirinto numa lista de corredores comprimidos.

corridorSmall :: [(Int, Piece)] -> [(Int, Piece)]

Junta os pares (1,peça) com o segundo membro igual somando o primeiro membro.

corridorNormal :: [Piece] -> [(Int, Piece)]

Transforma cada peça de um corredor (lista de peças) em pares (1,peça).