Tarefa4

Documentation

passTime :: Int -> State -> State

Introdução

Na Tarefa 4 introduziu-se a noção de tempo no jogo, permitindo assim a possibilidade de criar várias jogadas ao longo do tempo.

Objetivos e Desenvolvimento

Como objetivos desta tarefa tinhamos que efetuar a passagem do tempo num estado do jogo. Tentamos escrever o código o mais compacto possível e também a deixá-lo mais otimizado.

Visto que o jogo tinha apenas duas velocidades (pacman e fantasma vivo com velocidade 1 e fantasma morto com velocidade 0.5), decidimos dividir a função passTime em duas partes: quando o step é par e quando o step é ímpar. Com isto, quando o step é par todos os jogadores excutam uma jogada e quando um step é ímpar apenas os jogadores com velocidade igual a 1.

Discussão e Conlusão

Durante a execução da tarefa 4 tivemos algumas dificuldades na compreensão da função principal (passTime). No entanto após a total compreensão da função passTime, o seu código tornou-se relativamente fácil de escrever.

Em jeito de conclusão, a realização desta tarefa foi bastante simples, apesar dos seus pequenos contratempos inicialmente.

Recebe um step e o estado de jogo e altera o estado em uma iteração. (Função Principal)

playsEven :: [Player] -> State -> State

Recebe a lista de jogadores e o estado de jogo e excuta as jogadas para todos os jogadores.

playsOdd :: [Player] -> State -> State

Recebe a lista de jogadores e o estado de jogo e excuta as jogadas para todos os jogadores com velocidade igual a 1.

defaultDelayTime :: Integer