Tarefa5

Documentation

ghostPlay :: State -> [Play]

Introdução

Na tarefa 5 criamos a lógica necessária para a movimentação dos fantasmas de forma a que um usuário possa ter um desafio durante o jogo.

Objetivos e Desenvolvimento

Como principal objetivo desta tarefa tinhamos de implementar um comportamento para os fantasmas nos seus variados modos.

Existem apenas dois tipos de comportamentos possíveis sendo eles o chaseMode (modo de perseguição) e o scatterMode (modo de fuga).

Tanto no chaseMode como no scatterMode o fantasma redireciona-se sempre que se depara com uma parede de maneira a aproximar-se do pacman ou a movimentar-se para a sua direita.

Discussão e Conlusão

Com a concretização desta Tarefa temos um robô capaz de seguir o pacman com poucas dificuldades proporcionando assim alguma dificuldade ao usuário ou ao robô pacman desenvolvido na Tarefa seguinte.

Recebe um estado de jogo e devolve os movimentos de todos os fantasmas. (Função Principal)

idealChaseOri :: Player -> State -> Orientation

Recebe um fantasma e o estado de jogo e devolve a melhor orientação para perseguir o pacman (esta apenas muda de direção quando o fantasma embate numa parede).

dist :: Coords -> Coords -> Double

Calcula a distância entre dois pontos

minOri :: Double -> Orientation -> Double -> Orientation -> Orientation

Compara a distância de duas orientações e devolve a orientação com menor distância.

chaseMode :: State -> Int -> Play

Recebe o estado de jogo e o identificador do fantasma e devolve a melhor jogada de perseguição quando o fantasma embate numa parede.

rightOri :: Orientation -> Orientation

Muda de orientação num eixo de -90º.

leftOri :: Orientation -> Orientation

Muda de orientação num eixo de +90º.

idealScatterOri :: Player -> State -> Orientation

Recebe um fantasma e o estado de jogo e devolve uma orientação (em que o fantasma sempre que se depare com uma parede, este se mova para a sua direita caso seja possível).

scatterMode :: State -> Int -> Play

Recebe um estado e o identificador do fantasma e deolve uma jogada (em que o fantasma sempre que se depare com uma parede, este se mova para a sua direita caso seja possível).